Sorted all warnings

+++ lightcookietesting.h
@@ -12,14 +12,14 @@ inline SDL_Texture* addTextures(SDL_Renderer* renderer, vector<vector<int>> fogc
 	SDL_RenderClear(renderer);
 	paddingy = paddingx * (heightOfIlluminateMe / widthOfIlluminateMe);
 
-	for(int i = 0; i < fogcookies.size(); i++) {
-		for(int j = 0; j < fogcookies[0].size(); j++) {
+	for(long long unsigned int i = 0; i < fogcookies.size(); i++) {
+		for(long long unsigned int j = 0; j < fogcookies[0].size(); j++) {
 			if(fogcookies[i][j]) {
 				//render the lightspot
-				SDL_Rect dstrect = {i * ( (widthOfIlluminateMe ) / fogcookies.size()),
-				 					j * ( (heightOfIlluminateMe) /fogcookies[0].size()), 
-				 					(widthOfIlluminateMe + paddingx)/fogcookies.size(), 
-				 					(heightOfIlluminateMe + paddingy)/fogcookies[0].size()};
+				SDL_Rect dstrect = {(int)(i * ( (widthOfIlluminateMe ) / fogcookies.size())),
+				 					(int)(j * ( (heightOfIlluminateMe) /fogcookies[0].size())), 
+				 					(int)((widthOfIlluminateMe + paddingx)/fogcookies.size()), 
+				 					(int)((heightOfIlluminateMe + paddingy)/fogcookies[0].size())};
 				
 				SDL_RenderCopy(renderer, lightspot, NULL, &dstrect);
